app/requirements.txt
face_recognition
Pillow
flake8
dlib
rich
FastAPI
uvicorn
python-multipart
jinja2
celery
py-redis
pymongo
flower
boto3
python-magic
typing